requirements.txt
coloredlogs
cycler
humanfriendly
kiwisolver
matplotlib
numpy
opencv-contrib-python
opencv-python
Pillow
pynput
pyparsing
PyQt5
PyQt5-Qt5
PyQt5-sip
pyreadline
pyserial
python-dateutil
python-mpv
PyYAML
six
